Double/Float Tensor mismatch fixed

+++ RoboTrain.py
@@ -156,7 +156,7 @@ if __name__ == "__main__":
         bar = progressbar.ProgressBar(0,len(trainloader),redirect_stdout=False)
         for i, (images, labels) in enumerate(trainloader):
             if torch.cuda.is_available():
-                images = images.cuda()
+                images = images.float().cuda()
                 labels = labels.cuda()
 
             optimizer.zero_grad()
@@ -202,7 +202,7 @@ if __name__ == "__main__":
         bar = progressbar.ProgressBar(0, len(valloader), redirect_stdout=False)
         for i, (images, labels) in enumerate(valloader):
             if torch.cuda.is_available():
-                images = images.cuda()
+                images = images.float().cuda()
                 labels = labels.cuda()
 
             pred = model(images)
